transferred cart and order entity to JPA.

+++ src/main/java/com/mazanenko/petproject/bookshop/service/impl/CartServiceImpl.java
@@ -1,14 +1,12 @@
 package com.mazanenko.petproject.bookshop.service.impl;
 
-import com.mazanenko.petproject.bookshop.dao.CartDAO;
 import com.mazanenko.petproject.bookshop.entity.Book;
 import com.mazanenko.petproject.bookshop.entity.Cart;
-import com.mazanenko.petproject.bookshop.entity.Customer;
 import com.mazanenko.petproject.bookshop.entity.Order;
 import com.mazanenko.petproject.bookshop.entity.event.OrderEvent;
+import com.mazanenko.petproject.bookshop.repository.CartRepository;
 import com.mazanenko.petproject.bookshop.service.BookService;
 import com.mazanenko.petproject.bookshop.service.CartService;
-import com.mazanenko.petproject.bookshop.service.CustomerService;
 import com.mazanenko.petproject.bookshop.service.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
@@ -20,80 +18,100 @@ import java.sql.SQLException;
 @Service
 public class CartServiceImpl implements CartService {
 
-    private final CartDAO cartDAO;
+    private final CartRepository cartRepo;
     private final OrderService orderService;
-    private final CustomerService customerService;
     private final BookService bookService;
     private final ApplicationEventPublisher applicationEventPublisher;
 
     @Autowired
-    public CartServiceImpl(CartDAO cartDAO, OrderService orderService, CustomerService customerService,
+    public CartServiceImpl(CartRepository cartRepo, OrderService orderService,
                            BookService bookService, ApplicationEventPublisher applicationEventPublisher) {
-        this.cartDAO = cartDAO;
+        this.cartRepo = cartRepo;
         this.orderService = orderService;
-        this.customerService = customerService;
         this.bookService = bookService;
         this.applicationEventPublisher = applicationEventPublisher;
     }
 
     @Override
-    public Cart getCartById(int cartId) {
-        Cart cart = cartDAO.read(cartId);
-        if (cart != null) {
-            cart.setOrderList(orderService.readALLOrdersByCartId(cart.getId()));
+    public Cart getCartById(Long cartId) {
+        if (cartId == null) {
+            return null;
         }
-        return cart;
+        return cartRepo.findById(cartId).orElse(null);
     }
 
     @Override
-    public Cart getCartByCustomerId(int id) {
-        Cart cart = cartDAO.readByCustomerId(id);
-        if (cart != null) {
-            cart.setOrderList(orderService.readALLOrdersByCartId(cart.getId()));
+    public Cart getCartByCustomerId(Long customerId) {
+        if (customerId == null) {
+            return null;
         }
-        return cart;
+        return cartRepo.findByCustomer_Id(customerId).orElse(null);
     }
 
     @Override
     public Cart getCartByCustomerEmail(String email) {
-        Customer customer = customerService.getCustomerByEmail(email);
-        return getCartByCustomerId(customer.getId());
+        if (email == null) {
+            return null;
+        }
+        return cartRepo.findByCustomer_Email(email).orElse(null);
     }
 
     @Override
     @Transactional(rollbackFor = SQLException.class)
     public void addToCart(Order order) throws SQLException {
-        Order tempOrder = orderService.readOrderByCartIdAndProductId(order.getCartId(), order.getProductId());
+        if (order == null) {
+            return;
+        }
+        Order tempOrder = orderService.readOrderByCartIdAndProductId(order.getCart().getId(), order.getBook().getId());
 
         if (tempOrder == null) {
             orderService.createOrder(order);
         } else {
             orderService.incrementOrderQuantity(tempOrder.getId());
         }
-        bookService.decrementBookQuantity(order.getProductId());
+        bookService.decrementBookQuantity(order.getBook().getId());
     }
 
     @Override
-    public void addToCartByCustomerId(int customerId, int bookId) throws SQLException {
+    public void addToCartByCustomerId(Long customerId, Long bookId) throws SQLException {
+        if (customerId <= 0 || bookId <= 0) {
+            return;
+        }
+
         Cart cart = getCartByCustomerId(customerId);
-        Order order = new Order(cart.getId(), bookId, 1);
+        Book book = bookService.getBookById(bookId);
+        if (cart == null || book == null) {
+            return;
+        }
+
+        Order order = new Order(cart, book, 1);
         addToCart(order);
     }
 
     @Override
-    public void addToCartByCustomerEmail(String email, int bookId) throws SQLException {
+    public void addToCartByCustomerEmail(String email, Long bookId) throws SQLException {
+        if (email == null || bookId <= 0) {
+            return;
+        }
         Cart cart = getCartByCustomerEmail(email);
-        Order order = new Order(cart.getId(), bookId, 1);
+        Book book = bookService.getBookById(bookId);
+        if (cart == null || book == null) {
+            return;
+        }
+
+        Order order = new Order(cart, book, 1);
         addToCart(order);
     }
 
     @Override
     @Transactional(rollbackFor = SQLException.class)
-    public void incrementProduct(int productId, Cart cart) throws SQLException {
+    public void incrementProduct(Long productId, Cart cart) throws SQLException {
+        if (productId <= 0 || cart == null) {
+            return;
+        }
 
         for (Order order : cart.getOrderList()) {
-            if (order.getProductId() == productId) {
-
+            if (order.getBook().getId().equals(productId)) {
                 bookService.decrementBookQuantity(bookService.getBookById(productId).getId());
                 orderService.incrementOrderQuantity(order.getId());
             }
@@ -102,10 +120,13 @@ public class CartServiceImpl implements CartService {
 
     @Override
     @Transactional(rollbackFor = SQLException.class)
-    public void decrementProduct(int productId, Cart cart) {
+    public void decrementProduct(Long productId, Cart cart) {
+        if (productId <= 0 || cart == null) {
+            return;
+        }
 
         cart.getOrderList().forEach(order -> {
-            if (order.getProductId() == productId) {
+            if (order.getBook().getId().equals(productId)) {
                 orderService.decrementOrderQuantity(order.getId());
                 bookService.incrementBookQuantity(bookService.getBookById(productId).getId());
             }
@@ -113,17 +134,19 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public void updateOrderInCartById(int orderId, Order order) {
-        orderService.updateOrder(orderId, order);
-    }
-
-    @Override
     @Transactional(rollbackFor = SQLException.class)
-    public void deleteOrderFromCart(int productId, Cart cart) {
+    public void deleteOrderFromCart(Long productId, Cart cart) {
+        if (productId <= 0 || cart == null) {
+            return;
+        }
+
         Book book = bookService.getBookById(productId);
+        if (book == null) {
+            return;
+        }
 
         cart.getOrderList().forEach(order -> {
-            if (order.getProductId() == productId) {
+            if (order.getBook().getId().equals(productId)) {
                 int newQuantity = book.getAvailableQuantity() + order.getQuantity();
 
                 orderService.deleteOrder(order.getId());
@@ -133,21 +156,43 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public void deleteAllOrdersFromCart(int cartId) {
+    @Transactional(rollbackFor = Exception.class)
+    public void deleteAllOrdersFromCart(Long cartId) {
+        if (cartId <= 0) {
+            return;
+        }
+
+        Cart cart = getCartById(cartId);
+        if (cart == null) {
+            return;
+        }
+
+        cart.getOrderList().forEach(order -> deleteOrderFromCart(order.getBook().getId(), cart));
         orderService.deleteAllOrdersByCartId(cartId);
     }
 
     @Override
+    @Transactional
     public void makeAnOrderByCustomerEmail(String email) {
+        if (email == null) {
+            return;
+        }
+
         Cart cart = getCartByCustomerEmail(email);
 
-        deleteAllOrdersFromCart(cart.getId());
+        if (cart == null) {
+            return;
+        }
 
-        new Thread(() -> publishOrderEvent(cart)).start();
+        publishOrderEvent(cart);
+        orderService.deleteAllOrdersByCartId(cart.getId());
     }
 
 
     private void publishOrderEvent(Cart cart) {
+        if (cart == null) {
+            return;
+        }
         applicationEventPublisher.publishEvent(new OrderEvent(cart));
     }
 }
